fix: removed unused code
